+++ app_dart/lib/src/model/proto/protos.dart
@@ -6,6 +6,5 @@ export 'internal/build_status_response.pb.dart';
 export 'internal/commit.pb.dart';
 export 'internal/commit_status.pb.dart';
 export 'internal/key.pb.dart';
-export 'internal/scheduler.pb.dart';
 export 'internal/stage.pb.dart';
 export 'internal/task.pb.dart';
