+++ app_dart/lib/src/service/luci.dart
@@ -6,6 +6,7 @@ import 'dart:async';
 import 'dart:math';
 
 import 'package:appengine/appengine.dart';
+import 'package:cocoon_scheduler/scheduler.dart';
 import 'package:github/github.dart';
 import 'package:json_annotation/json_annotation.dart';
 import 'package:meta/meta.dart';
@@ -13,7 +14,6 @@ import 'package:retry/retry.dart';
 
 import '../model/appengine/task.dart';
 import '../model/luci/buildbucket.dart';
-import '../model/proto/internal/scheduler.pb.dart';
 import '../request_handling/api_request_handler.dart';
 import 'buildbucket.dart';
 import 'config.dart';
