+++ app_dart/pubspec.yaml
@@ -12,6 +12,8 @@ environment:
 
 dependencies:
   appengine: ^0.11.0
+  cocoon_scheduler:
+    path: ../scheduler
   collection: ^1.14.11
   corsac_jwt: ^0.2.2
   crypto: ^2.0.6
@@ -31,7 +33,6 @@ dependencies:
   neat_cache: ^1.0.1
   protobuf: ^1.0.0
   truncate: ^2.1.2
-  yaml: ^2.1.16
 
 dev_dependencies:
   build_runner: ^1.0.0
