+++ scheduler/README.md
@@ -1,0 +1,3 @@
+# Scheduler
+
+Logic for Flutter infrastructure tasks being scheduled on physical devices.
