+++ scheduler/pubspec.yaml
@@ -1,0 +1,15 @@
+name: cocoon_scheduler
+description: Flutter infrastructure scheduler
+version: 1.0.0
+publish_to: none
+
+environment:
+  sdk: '>=2.10.0 <3.0.0'
+
+dependencies:
+  protobuf: ^1.0.0
+  yaml: ^2.1.16
+
+dev_dependencies:
+  pedantic: ^1.10.0
+  test: ^1.16.0
