+++ tests.yaml
@@ -23,6 +23,9 @@ tasks:
   - task: app_dart
     script: test_utilities/bin/dart_test_runner.sh
 
+  - task: scheduler
+    script: test_utilities/bin/dart_test_runner.sh
+
   - task: app_flutter
     script: test_utilities/bin/flutter_test_runner.sh
